Configured validations
I Configured the validation build process
validations were build from the types and interfaces using the
runtime-typescript-checker library.

The validation can be build using the following command:
  npm run build-validation
or
  npm run build-validation:force
to force a new build even if the validation files are already present.

the process generates a json file for each type and interface in the
src/validation folder.
and also a file with all the validations in the typings folder where it picks up the
typings from. thought i'm looking for a better way to do this
and place theme in a single folder upon build. It seems that the
runtime-typescript-checker library does not support this at the moment.