Wrote the OutBound Class
The outbound class houses all the methods for sending out messages.

It can only handle media messages and with either url or media ids but
uploading media file to the graph client is not ready yet.

I will have to go back to facebook's documentation for for acceptable
media types and size limitaions before building that into the class